+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.component.ts
@@ -1416,6 +1416,29 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   }
 
   /**
+   * Return an array containing all of the currently modified rows, the type of modification
+   * and the cells that are dirty and the data.
+   * @returns An keyed object showing the dirty row info.
+   */
+  getModifiedRows(): SohoDataGridModifiedRows {
+    return this.ngZone.runOutsideAngular(() => {
+      return this.datagrid.getModifiedRows();
+    });
+  }
+
+  /**
+   * Set a cell to dirty and add the dirty icon visually.
+   * @param row The row index
+   * @param cell The cell index
+   * @param toggle True to set it and false to remove it
+   */
+  setDirtyIndicator(row: number, cell: number, toggle: boolean): void {
+    this.ngZone.runOutsideAngular(() => {
+      this.datagrid.setDirtyIndicator(row, cell, toggle);
+    });
+  }
+
+  /**
    * Removes all selected rows
    */
   removeSelected() {
@@ -1732,6 +1755,25 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
     this.ngZone.run(() => {
       this.expandrow.next(event);
     });
+  }
+
+  /**
+   * Event fired after a child row has been expanded.
+   * @param idProperty string id
+   */
+  @Input() set onBeforeSelect(beforeSelectFunction: SohoDataGridBeforeSelectFunction) {
+    this._gridOptions.onBeforeSelect = beforeSelectFunction;
+    if (this.datagrid) {
+      this.datagrid.settings.onBeforeSelect = beforeSelectFunction;
+      this.markForRefresh('onBeforeSelect', RefreshHintFlags.Rebuild);
+    }
+  }
+
+  get onBeforeSelect(): SohoDataGridBeforeSelectFunction {
+    if (this.datagrid) {
+      return this.datagrid.settings.onBeforeSelect;
+    }
+    return this._gridOptions.onBeforeSelect;
   }
 
   /**