+++ src/app/app.routes.ts
@@ -111,6 +111,7 @@ import { ModalDialogDemoComponent } from './modal-dialog/modal-dialog.demo';
 import { MonthViewDemoComponent } from './monthview/monthview.demo';
 import { NotificationDemoComponent } from './notification/notification.demo';
 import { PagerStandaloneDemoComponent } from './pager/pager-standalone.demo';
+import { PersonalizeColorApiDemoComponent } from './personalize/personalize-color-api.demo';
 import { PieDemoComponent } from './pie/pie.demo';
 import { PopDownDemoComponent } from './popdown/popdown.demo';
 import { PopupMenuDemoComponent } from './popupmenu/popupmenu.demo';
@@ -149,6 +150,7 @@ import { ToolbarAllIconsDemoComponent } from './toolbar/toolbar-all-icons.demo';
 import { ToolbarBasicDemoComponent } from './toolbar/toolbar-basic.demo';
 import { ToolbarDataDrivenDemoComponent } from './toolbar/toolbar-datadriven.demo';
 import { ToolbarFlexBasicDemoComponent } from './toolbar-flex/toolbar-flex-basic.demo';
+import { ToolbarFlexDatagridDemoComponent } from './toolbar-flex/toolbar-flex-datagrid.demo';
 import { ToolbarFlexMoreActionsAjaxDemoComponent } from './toolbar-flex/toolbar-flex-more-actions-ajax.demo';
 import { ToolbarFlexSearchfieldDemoComponent } from './toolbar-flex/toolbar-flex-searchfield.demo';
 import { ToolbarMoreActionsAjaxDemoComponent } from './toolbar/toolbar-more-actions-ajax.demo';
@@ -282,6 +284,7 @@ export const routes: Routes = [
   { path: 'month-view',                             component: MonthViewDemoComponent },
   { path: 'notification',                           component: NotificationDemoComponent },
   { path: 'pager-standalone',                       component: PagerStandaloneDemoComponent },
+  { path: 'personalize-color-api',                  component: PersonalizeColorApiDemoComponent },
   { path: 'pie',                                    component: PieDemoComponent },
   { path: 'popdown',                                component: PopDownDemoComponent},
   { path: 'popupmenu',                              component: PopupMenuDemoComponent },
@@ -323,6 +326,7 @@ export const routes: Routes = [
   { path: 'toolbar-state',                          component: ToolbarStateComponent },
   { path: 'toolbar-right-aligned',                  component: ToolbarRightAlignedDemoComponent },
   { path: 'toolbar-flex-basic',                     component: ToolbarFlexBasicDemoComponent },
+  { path: 'toolbar-flex-datagrid',                  component: ToolbarFlexDatagridDemoComponent },
   { path: 'toolbar-flex-more-actions-ajax',         component: ToolbarFlexMoreActionsAjaxDemoComponent },
   { path: 'toolbar-flex-searchfield',               component: ToolbarFlexSearchfieldDemoComponent },
   { path: 'tooltip',                                component: TooltipDemoComponent },